refactor component modules
